Warning消し
git-svn-id: https://svn.osdn.net/svnroot/ttssh2/trunk@3691 f5f01b69-1e22-0410-acbf-894ab4bd6246

+++ teraterm/teraterm/clipboar.c
@@ -165,7 +165,7 @@ void CBSend()
 	}
 
 	now = GetTickCount();
-	if (now - lastcr < ts.PasteDelayPerLine) {
+	if (now - lastcr < (DWORD)ts.PasteDelayPerLine) {
 		return;
 	}
